refactor(desktop): split v2 FilesTab into bridge hook + sub-components
FilesTab.tsx grew to ~1080 lines after the @pierre/trees swap. Split out
three concerns:

- `hooks/useFilesTabBridge`: owns the lazy-load plumbing (knownPaths,
  loadedDirs, pendingCreates, fetchDir, doRefresh) plus the three side
  effects (initial load, subscribe-driven lazy expand, fs:events sync).
  This is the Pierre↔backend boundary; isolating it makes the contract
  clear and the rest of FilesTab much easier to read.
- `components/{RowContextMenu,FileMenuItems,FolderMenuItems,PathActions}`:
  pure rendering for the right-click menu, one component per file per
  AGENTS.md.
- `utils/treePath`: pure path helpers (toRel/toAbs/parentRel/basename/
  asDirectoryHandle), unit-testable in isolation.

FilesTab.tsx is now ~670 lines and reads as orchestration: model
construction, ref-callback wiring for Pierre's stable callbacks, the
inline create/rename flow, and click-intent intercept.
