+++ apps/desktop/src/renderer/routes/_authenticated/_dashboard/v2-workspace/$workspaceId/components/WorkspaceSidebar/components/FilesTab/components/FolderMenuItems/FolderMenuItems.tsx
@@ -1,0 +1,43 @@
+import {
+	DropdownMenuItem,
+	DropdownMenuSeparator,
+} from "@superset/ui/dropdown-menu";
+import { PathActions } from "../PathActions";
+
+interface FolderMenuItemsProps {
+	absolutePath: string;
+	relativePath: string;
+	onNewFile: () => void;
+	onNewFolder: () => void;
+	onRename: () => void;
+	onDelete: () => void;
+}
+
+export function FolderMenuItems({
+	absolutePath,
+	relativePath,
+	onNewFile,
+	onNewFolder,
+	onRename,
+	onDelete,
+}: FolderMenuItemsProps) {
+	return (
+		<>
+			<DropdownMenuItem onSelect={() => setTimeout(onNewFile, 0)}>
+				New File...
+			</DropdownMenuItem>
+			<DropdownMenuItem onSelect={() => setTimeout(onNewFolder, 0)}>
+				New Folder...
+			</DropdownMenuItem>
+			<DropdownMenuSeparator />
+			<PathActions absolutePath={absolutePath} relativePath={relativePath} />
+			<DropdownMenuSeparator />
+			<DropdownMenuItem onSelect={() => setTimeout(onRename, 0)}>
+				Rename...
+			</DropdownMenuItem>
+			<DropdownMenuItem variant="destructive" onSelect={onDelete}>
+				Delete
+			</DropdownMenuItem>
+		</>
+	);
+}
