+++ apps/desktop/src/renderer/routes/_authenticated/_dashboard/v2-workspace/$workspaceId/components/WorkspaceSidebar/components/FilesTab/hooks/useFilesTabBridge/index.ts
@@ -1,0 +1,2 @@
+export type { FilesTabBridge } from "./useFilesTabBridge";
+export { useFilesTabBridge } from "./useFilesTabBridge";
